requirements.txt
nonebot2[httpx,fastapi,websockets]>=2.4.2
nonebot-adapter-qq>=1.6.3
